3.1.12 przeorganizowanie projektu (*)
