--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,32 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class Node {
-
-    private int value;
-    private Node prev;
-    private Node next;
-
-    public Node(int i) {
-        value = i;
-    }
-
-    int getValue() {
-        return value;
-    }
-
-    Node getPrev() {
-        return prev;
-    }
-
-    void setPrev(final Node prev) {
-        this.prev = prev;
-    }
-
-    Node getNext() {
-        return next;
-    }
-
-    void setNext(final Node next) {
-        this.next = next;
-    }
-}
